Refactored and packaged code
