+++ test.py
@@ -1,0 +1,3 @@
+from tischlre_diffusion2d.diffusion2d import solve
+
+solve(dx = 0.1, dy = 0.1, D = 4)
